# perf: comparison benchmarks — remove loop overhead, guard invariants; refresh docs