标题：perf: comparison benchmarks — remove loop overhead, guard invariants; refresh docs

动机
- 目标场景：对比基准（gint vs ClickHouse vs Boost）在某些用例中含有循环内拷贝/构造与潜在常量折叠，导致测得时间包含非算子开销或被编译器优化掉，影响公正性与可复现性。
- 基线：Add_FullCarry/Sub_FullBorrow 等“最坏路径”可能被当作常量表达式外提；Mul 与 Div 的部分用例在循环内构造操作数，增加非算子成本。

优化思路
- 算法/路径：不改动库实现，仅修正基准方法学。
- 具体措施：
  - 预生成数据集，循环中仅取引用并计算；移除 per-iteration 的拷贝/构造/赋值。
  - 对循环不变操作数使用非 const 版本 `DoNotOptimize`，并从静态数据集加载，避免常量折叠/外提。
  - 统一对表达式本身调用 `DoNotOptimize(expr)`，减少临时与赋值开销。
- 位宽专用化：不涉及。
- 可维护性：bench 代码更简洁；文档更新了方法学说明与新结果。

正确性
- 语义不变：未改动库代码与行为；仅修改基准与文档。
- 覆盖测试：`make test` 132 个用例全部通过（AppleClang 17, macOS）。

基准对比（必填）
- 环境：Apple Silicon（12 cores），AppleClang 17，`-O3 -DNDEBUG`，Google Benchmark；`--benchmark_min_time=0.2s`。
- 方法：
  - 核心算子：`make bench BENCH_ARGS="--benchmark_min_time=0.2s"`
  - 对比矩阵：`make bench-compare-full BENCH_ARGS="--benchmark_min_time=0.2s"`
- 关键用例修正前/后（ns/op，示例）：
  - Add/FullCarry (gint)：0.47 → 1.17（修正后避免常量折叠）
  - Sub/FullBorrow (gint)：0.98 → 1.67（修正后避免常量折叠）
  - 其它代表性用例（如 Div/Pow2Divisor）相对排序与数量级保持稳定，但更“干净”地反映算法开销。
- 全量结果：已写入 `docs/BENCHMARKS.md` 与 `README.md` 的表格（见本 PR 文件差异）。
- 稳定性：固定种子，预生成 256 组样本，每用例循环轮询；`--benchmark_min_time=0.2s`；macOS 上线程亲和性警告不影响相对比较。

回归与风险
- 平台差异：基准代码对 Clang/GCC 均兼容；不依赖新特性。
- 代码尺寸与可读性：bench 更清晰；文档同步更新方法学说明。

清单（提交前请逐项勾选）
- [x] 分支名为英文（PR 指南要求）
- [x] 已运行 clang-format（`.clang-format`）
- [x] 新增/更新了 `tests/` 覆盖关键路径，`make test` 通过（本 PR 未改库实现，测试未变）
- [x] 已提供“变更前/后”基准数据（见上）
- [x] 若更改基准用例，已更新 `bench/` 与 `docs/BENCHMARKS.md`
- [x] 保持 C++11 兼容与 Header-only 特性
- [x] 保持严格位宽与原生类型互操作语义不变